Added: Missing Padding to CharacterParameters

+++ Sewer56.SonicRiders/Structures/Enums/Gender.cs
@@ -1,6 +1,6 @@
 ﻿namespace Sewer56.SonicRiders.Structures.Enums
 {
-    public enum Gender : byte
+    public enum Gender : int
     {
         Male,
         Female
